Add files via upload
CFFFP_TopFrontWheelSupport.gcode is a bracket part that also had layer skipping failure around the critical area where the screw holes are. This part takes only 1hr30min to print.
CFFFP_WingMountBottom.gcode is the one that failed miserably to print it prints a little bit then it starts to skip.

WingMountBottom.STL is the 3D model iof the part that failed miserably to print in xvico 3X in STL format.